requirements.txt
alembic==1.13.2
attrs==24.2.0
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
flask-marshmallow==1.2.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
marshmallow==3.22.0
marshmallow-sqlalchemy==1.1.0
packaging==24.1
pluggy==1.5.0
py==1.11.0
PyJWT==2.9.0
pytest==7.1.2
python-dotenv==1.0.1
setuptools==74.1.2
six==1.16.0
SQLAlchemy==2.0.34
tomli==2.0.1
typing_extensions==4.12.2
Werkzeug==3.0.4
